Components (Overlay & RTS): Fix compiling issues regarding mismatch and ambiguous between float and double.

+++ Components/Overlay/src/OgreFont.cpp
@@ -540,7 +540,7 @@ namespace Ogre
                            (Real)m / (Real)finalHeight,                  // v1
                            (Real)(l + width) / (Real)finalWidth,         // u2
                            (m + max_height) / (Real)finalHeight); // v2
-                this->setGlyphInfo({cp, uvs, textureAspect * uvs.width() / uvs.height(),
+                this->setGlyphInfo({cp, uvs, float(textureAspect * uvs.width() / uvs.height()),
                                     float(x_bearing) / max_height, float(advance) / max_height});
 
                 // Advance a column
